refactor(modules): moved module class definition
