+++ docs/reference/config.js
@@ -1,1 +1,1 @@
-exports.fileTree=[{"name":"client","path":"/client","fullPath":"./docs/reference/client"},{"name":"_index","path":"/_index","fullPath":"./docs/reference/_index"},{"name":"modules","children":[{"name":"auth","children":[{"name":"email","children":[{"name":"_index","path":"/_index","fullPath":"modules/auth/email/_index"}]},{"name":"_index","path":"/_index","fullPath":"modules/auth/_index"}]},{"name":"core","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/_index"},{"name":"role","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/role/_index"}]},{"name":"user","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/user/_index"}]},{"name":"webservice","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/webservice/_index"}]}]},{"name":"_index","path":"/_index","fullPath":"modules/_index"},{"name":"shared","children":[]}]}];exports.sidebarTree = (title = 'Mainpage') => ({"/reference/":[{"title":"API Reference","collapsable":false,"children":[["",""+title+""],"client","_index"]},{"title":"modules","collapsable":false,"children":["modules/auth/email/_index","modules/auth/_index","modules/core/_index","modules/core/role/_index","modules/core/user/_index","modules/core/webservice/_index","modules/_index"]}]});
+exports.fileTree=[{"name":"client","path":"/client","fullPath":"./docs/reference/client"},{"name":"_index","path":"/_index","fullPath":"./docs/reference/_index"},{"name":"module","path":"/module","fullPath":"./docs/reference/module"},{"name":"modules","children":[{"name":"auth","children":[{"name":"email","children":[{"name":"_index","path":"/_index","fullPath":"modules/auth/email/_index"}]},{"name":"_index","path":"/_index","fullPath":"modules/auth/_index"}]},{"name":"core","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/_index"},{"name":"role","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/role/_index"}]},{"name":"user","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/user/_index"}]},{"name":"webservice","children":[{"name":"_index","path":"/_index","fullPath":"modules/core/webservice/_index"}]}]},{"name":"shared","children":[]}]}];exports.sidebarTree = (title = 'Mainpage') => ({"/reference/":[{"title":"API Reference","collapsable":false,"children":[["",""+title+""],"client","_index","module"]},{"title":"modules","collapsable":false,"children":["modules/auth/email/_index","modules/auth/_index","modules/core/_index","modules/core/role/_index","modules/core/user/_index","modules/core/webservice/_index"]}]});
