+++ src/modules/auth/index.ts
@@ -4,7 +4,7 @@
  * title: auth
  * ---
  */
-import Module from "..";
+import Module from "../../module";
 import AuthEmailModule from "./email";
 import { Client } from "../../client";
